Create serialization module and update sphinx docs
AnyMetadataDeserializer and PayloadDeserializer moved to different
module, `models.serialization`. To avoid circular imports these
deserializers are imported inside default_deserializer of Metadata.
pylint werning are disabled for these imports.

Some documentation correction has been done, and sphinx docs are
enabled for new Metadata classes and deserializers.

in-toto's DSSE envelope `payload_type` is now stored under a
constant variable.

create_envelope() method of Signed now utilizes in-toto version
of DSSE envelope rather than securesystemslib version.
